requirements.txt
django
djangorestframework
RPi.GPIO